Fix Lat and Long Datatype

+++ item/models.py
@@ -5,8 +5,8 @@ from django.db import models
 class EssentialItem(models.Model):
 	name = models.CharField(max_length=120)
 	quantity = models.IntegerField()
-	lat = models.IntegerField()
-	long = models.IntegerField()
+	lat = models.FloatField()
+	long = models.FloatField()
 
 	def __str__(self):
 		return self.name
